Saturday afternoon fever

+++ ft_strlcpy.c
@@ -6,7 +6,7 @@
 /*   By: samartin <samartin@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/14 14:06:24 by samartin          #+#    #+#             */
-/*   Updated: 2022/09/17 19:44:45 by samartin         ###   ########.fr       */
+/*   Updated: 2022/09/23 16:44:37 by samartin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@ size_t	ft_strlcpy(char *dst, const char *src, size_t dstsize)
 	size_t	pos;
 
 	pos = 0;
-	while ((src[pos] != 0) && (pos < (dstsize - 1)))
+	while ((src[pos] != 0) && ((int)pos < (int)(dstsize - 1)))
 	{
 		dst[pos] = src[pos];
 		pos++;
